MOved library autofocusutils to imageprocessing.  Seems more logical today.
